VERSION_3_12032023

+++ lib/view/equipos/editEquipoJugador.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_image_compress/flutter_image_compress.dart';
@@ -29,9 +30,9 @@ class EditEquipoJugador extends StatefulWidget {
 class _EditEquipoJugadorState extends State<EditEquipoJugador> {
   final _formKey = GlobalKey<FormState>();
   final _nombre = TextEditingController();
-  PickedFile _imageFile;
-  String imageUrl;
-  File _image;
+
+   String? imageUrl;
+   XFile? _imageFileList;
   bool insertar=false;
   dynamic _pickImageError;
   final ImagePicker _picker = ImagePicker();
@@ -54,78 +55,29 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
     super.initState();
   }
 
-  Future getImage() async {
-    var image = await ImagePicker.pickImage(
-      source: ImageSource.gallery,imageQuality: 20,
-        maxHeight:  100 , maxWidth: 100);
 
-    setState(() {
-      _image = image;
-      print('Image Path $_image');
-    });
-  }
 
-  void _showPicker(context) {
-    showModalBottomSheet(
-        context: context,
-        builder: (BuildContext bc) {
-          return SafeArea(
-            child: Container(
-              child: new Wrap(
-                children: <Widget>[
-                  new ListTile(
-                      leading: new Icon(Icons.photo_library),
-                      title: new Text('Photo Library'),
-                      onTap: () {
-                        _imgFromGallery();
-                        Navigator.of(context).pop();
-                      }),
-                  new ListTile(
-                    leading: new Icon(Icons.photo_camera),
-                    title: new Text('Camera'),
-                    onTap: () {
-                      _imgFromCamera();
-                      Navigator.of(context).pop();
-                    },
-                  ),
-                ],
-              ),
-            ),
-          );
-        });
-  }
+  Future<void> _onImageButtonPressed(ImageSource source,
+      {BuildContext? context, bool isMultiImage = false}) async {
+          try{
+              final XFile? pickedFile = await _picker.pickImage(
+                source: ImageSource.gallery,
+                maxWidth: 100,
+                maxHeight: 100,
+                imageQuality: 100,
+              );
+              setState(() {
+                _imageFileList= pickedFile!;
 
-  _imgFromCamera() async {
-    try {
-      final pickedFile = await _picker.getImage(
-        source: ImageSource.camera,
-        imageQuality: 20,
-      );
-      setState(() {
-        _imageFile = pickedFile;
-      });
-    } catch (e) {
-      setState(() {
-        _pickImageError = e;
-      });
+              });
+            } catch (e) {
+              setState(() {
+                print("EEEERRRRRROOOOOORR:${e.toString()}");
+                _pickImageError = e;
+              });
+            }
     }
-  }
 
-  _imgFromGallery() async {
-    try {
-      final pickedFile = await _picker.getImage(
-        source: ImageSource.gallery,
-        imageQuality: 20,
-      );
-      setState(() {
-        _imageFile = pickedFile;
-      });
-    } catch (e) {
-      setState(() {
-        _pickImageError = e;
-      });
-    }
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -142,7 +94,7 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
         icon: Icon(Icons.person),
       ),
       validator: (value) {
-        if (value.isEmpty) {
+        if (value!.isEmpty) {
           return 'Incorrecto';
         }
         return null;
@@ -159,7 +111,7 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
         Center(
           child: GestureDetector(
             onTap: () {
-              _showPicker(context);
+              _onImageButtonPressed(ImageSource.camera, context: context);
             },
             child: CircleAvatar(
               radius: 55  ,
@@ -167,38 +119,36 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
               child:  new SizedBox(
                     width: 120.0,
                     height: 120.0,
-                    child: _imageFile != null
+                    // ignore: unnecessary_null_comparison
+
+                    child:
+                    !kIsWeb?
+                     _imageFileList != null
                     //https://firebasestorage.googleapis.com/v0/b/iaclub.appspot.com/o/
-                        ? Image.file(File(_imageFile.path), fit: BoxFit.fill)
+                        ? Image.file(File(_imageFileList!.path), fit: BoxFit.fill)
                         : Image.network("https://firebasestorage.googleapis.com/v0/b/iafootfeel.appspot.com/"
                         "o/clubes"
                         "%2F${widget.equipo.equipo.replaceAll("Ñ", "N").replaceAll("ñ", "n")}.png"
                         "?alt=media",
                             fit: BoxFit.cover,height: 90,
-                          )),
+                            )
+                  :_imageFileList != null
+                    //https://firebasestorage.googleapis.com/v0/b/iaclub.appspot.com/o/
+                    ? Image.network(_imageFileList!.path, fit: BoxFit.fill)
+                        : Image.network("https://firebasestorage.googleapis.com/v0/b/iafootfeel.appspot.com/"
+                    "o/clubes"
+                    "%2F${widget.equipo.equipo.replaceAll("Ñ", "N").replaceAll("ñ", "n")}.png"
+                    "?alt=media",
+                    fit: BoxFit.cover,height: 90,
+                    ))
+              ,
             ),
           ),
         )
       ],
     );
 
-    Future<DateTime> getDate() {
-      // Imagine that this function is
-      // more complex and slow.
-      return showDatePicker(
-        context: context,
-        initialDate: DateTime.now(),
-        firstDate: DateTime(1980),
-        lastDate: DateTime(2050),
-        locale : const Locale("es","ES"),
-        builder: (BuildContext context, Widget child) {
-          return Theme(
-            data: ThemeData.dark(),
-            child: child,
-          );
-        },
-      );
-    }
+
 
     ListView body = ListView(
       padding: EdgeInsets.all(0),
@@ -237,9 +187,9 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
               child: IconButton(
                   icon: Icon(Icons.save_outlined,size: 30, color: Colors.blue,),
                   onPressed: () async {
-                    if (_formKey.currentState.validate()) {
-                      _formKey.currentState.save();
-                      widget.equipo.equipo = inputNombre.controller.text;
+                    if (_formKey.currentState!.validate()) {
+                      _formKey.currentState!.save();
+                      widget.equipo.equipo = inputNombre.controller!.text;
                        if(insertar) {
                           await equipoProvider.addEquipo(widget.pais,
                             widget.equipo);
@@ -247,7 +197,7 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
                         await equipoProvider.updateEquipo(widget.pais,
                             widget.equipo);
                       }
-                      if(_imageFile!=null)
+                      if(_imageFileList!=null)
                         await uploadFile();
                       Navigator.pop(context);
                     }
@@ -271,38 +221,57 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
   Future uploadFile() async {
     print("UPLLOAD");
     final _storage = FirebaseStorage.instance;
-    final _picker = ImagePicker();
-    await Permission.photos.request();
-    var permissionStatus = await Permission.photos.status;
+
+    //   await Permission.photos.request();
 
    // if (permissionStatus.isGranted) {
       //Select Image
-      print("STORAGE:${_imageFile.path}");
-      var file = File(_imageFile.path);
+      print("STORAGE:${_imageFileList!.path}");
+      var file = File(_imageFileList!.path);
     setState(() {});
-      print("STORAGE:${_imageFile.path}");
 
-      if (_imageFile != null) {
-        final tempDir = await getTemporaryDirectory();
-        final path = tempDir.path;
+      if (_imageFileList != null) {
+        print("STORAGE:${_imageFileList!.path}");
 
-        Im.Image image = Im.decodeImage(file.readAsBytesSync());
-        image= Im.copyResize(image,width: 500,);
-        // choose the size here, it will maintain aspect ratio
-        var compressedImage = new File('$path/img_copy.png')..writeAsBytesSync(Im.encodePng(image,level: 4));
-
-
+        var compressedImage;
+        if (!kIsWeb) {
+          print("kIsWeb:${_imageFileList!.path}");
+          final tempDir = await getTemporaryDirectory();
+          final path = tempDir.path;
+          Im.Image? image = Im.decodeImage(file.readAsBytesSync());
+          image = Im.copyResize(image!, width: 100,);
+          // choose the size here, it will maintain aspect ratio
+           compressedImage = new File('$path/img_copy.png')
+            ..writeAsBytesSync(Im.encodePng(image, level: 4));
+        }
         //Upload to Firebase
-        var snapshot =
-            await _storage.ref().child(
+        print("ANTES:${_imageFileList!.path}");
+
+        var snapshot;
+          if(kIsWeb==false) {
+            print("IOS:${_imageFileList!.path}");
+
+            snapshot= await _storage.ref().child(
                 "clubes"
                     "/${widget.equipo.equipo.replaceAll("Ñ", "N").
-                replaceAll("ñ", "n")}.png").putFile(compressedImage, SettableMetadata(contentType: 'image/png',));
-        var downloadUrl = await snapshot.ref.getDownloadURL();
+                replaceAll("ñ", "n")}.png").putFile(
+                compressedImage, SettableMetadata(contentType: 'image/png',));
+          }else {
+            print("WEB:${_imageFileList!.path}");
+            Uint8List aux=await _imageFileList!.readAsBytes();
+            snapshot=await _storage.ref().child(
+                "clubes"
+                    "/${widget.equipo.equipo.replaceAll("Ñ", "N").
+                replaceAll("ñ", "n")}.png").putData(aux);
+          }
 
+        var downloadUrl = await snapshot.ref.getDownloadURL();
+        print("DOWN:${downloadUrl}");
         setState(() {
           imageUrl = downloadUrl;
-          compressedImage.delete();
+          if(kIsWeb==false) {
+            compressedImage.delete();
+          }
         });
       } else {
         print('No Path Received');
@@ -312,19 +281,4 @@ class _EditEquipoJugadorState extends State<EditEquipoJugador> {
     }*/
   }
 
-
-  Future<File> testCompressAndGetFile(File file, String targetPath) async {
-    print("testCompressAndGetFile");
-    final result = await FlutterImageCompress.compressAndGetFile(
-      file.absolute.path,
-      targetPath,
-      quality: 90,
-      minWidth: 100,
-      minHeight: 100,
-      rotate: 0,
-    );
-
-
-    return result;
-  }
 }
